Added the following: - Error Handling, Validations - Response Objects mapping - Documentation

+++ Sorted.Domain/Rainfall/ServiceResponseObject.cs
@@ -23,6 +23,6 @@
         public string id { get; set; }
         public DateTime dateTime { get; set; }
         public string measure { get; set; }
-        public float value { get; set; }
+        public decimal value { get; set; }
     }
 }
